Clean up test data generation
