+++ tests/conftest.py
@@ -2,7 +2,10 @@ import os
 
 import pytest
 
-DATA_DIR_NAME = "data"
+ALMANAC_DIR_NAME = "almanac"
+SMALL_SKY_DIR_NAME = "small_sky"
+SMALL_SKY_ORDER1_DIR_NAME = "small_sky_order1"
+
 
 TEST_DIR = os.path.dirname(__file__)
 SMALL_SKY_DIR_NAME = "small_sky"
@@ -49,3 +52,28 @@ def local_data_dir():
 @pytest.fixture
 def small_sky_dir_local(local_data_dir):
     return os.path.join(local_data_dir, SMALL_SKY_DIR_NAME)
+
+
+@pytest.fixture
+def tmp_dir_cloud(example_cloud_path):
+    return os.path.join(example_cloud_path, "tmp")
+
+
+@pytest.fixture
+def test_data_dir_cloud(example_cloud_path):
+    return os.path.join(example_cloud_path, "data")
+
+
+@pytest.fixture
+def almanac_dir_cloud(test_data_dir_cloud):
+    return os.path.join(test_data_dir_cloud, ALMANAC_DIR_NAME)
+
+
+@pytest.fixture
+def small_sky_dir_cloud(test_data_dir_cloud):
+    return os.path.join(test_data_dir_cloud, SMALL_SKY_DIR_NAME)
+
+
+@pytest.fixture
+def small_sky_order1_dir_cloud(test_data_dir_cloud):
+    return os.path.join(test_data_dir_cloud, SMALL_SKY_ORDER1_DIR_NAME)
