+++ tests/data/small_sky/catalog_info.json
@@ -1,12 +1,8 @@
 {
     "catalog_name": "small_sky",
-    "catalog_type": "source",
-    "version": "0.0.1",
-    "generation_date": "2022.12.20",
+    "catalog_type": "object",
+    "total_rows": 131,
     "epoch": "J2000",
-    "ra_kw": "ra",
-    "dec_kw": "dec",
-    "id_kw": "id",
-    "total_objects": 131,
-    "pixel_threshold": 1000000
+    "ra_column": "ra",
+    "dec_column": "dec"
 }
