+++ tests/data/small_sky/partition_info.csv
@@ -1,2 +1,2 @@
-Norder,Dir,Npix,num_rows
-0,0,11,131
+Norder,Npix,Dir
+0,11,0
