+++ tests/data/small_sky_order1/partition_info.csv
@@ -1,5 +1,5 @@
-Norder,Dir,Npix,num_rows
-1,0,44,42
-1,0,45,29
-1,0,46,42
-1,0,47,18
+Norder,Npix,Dir
+1,44,0
+1,45,0
+1,46,0
+1,47,0
