+++ tests/data/small_sky_order1/provenance_info.json
@@ -1,0 +1,55 @@
+{
+    "catalog_name": "small_sky_order1",
+    "catalog_type": "object",
+    "total_rows": 131,
+    "epoch": "J2000",
+    "ra_column": "ra",
+    "dec_column": "dec",
+    "version": "0.2.9.dev2+g014342d",
+    "generation_date": "2024.03.18",
+    "tool_args": {
+        "tool_name": "hipscat_import",
+        "version": "0.2.6.dev6+gf95440a",
+        "runtime_args": {
+            "catalog_name": "small_sky_order1",
+            "output_path": ".",
+            "output_artifact_name": "small_sky_order1",
+            "tmp_dir": "/tmp/user/11115/tmpw6p6hl9p",
+            "overwrite": true,
+            "dask_tmp": "",
+            "dask_n_workers": 1,
+            "dask_threads_per_worker": 1,
+            "catalog_path": "./small_sky_order1",
+            "tmp_path": "/tmp/user/11115/tmpw6p6hl9p/small_sky_order1/intermediate",
+            "epoch": "J2000",
+            "catalog_type": "object",
+            "input_path": "small_sky_parts",
+            "input_paths": [
+                "file:///home/delucchi/git/hipscat-cloudtests/tests/data/small_sky_parts/catalog_00_of_05.csv",
+                "file:///home/delucchi/git/hipscat-cloudtests/tests/data/small_sky_parts/catalog_01_of_05.csv",
+                "file:///home/delucchi/git/hipscat-cloudtests/tests/data/small_sky_parts/catalog_02_of_05.csv",
+                "file:///home/delucchi/git/hipscat-cloudtests/tests/data/small_sky_parts/catalog_03_of_05.csv",
+                "file:///home/delucchi/git/hipscat-cloudtests/tests/data/small_sky_parts/catalog_04_of_05.csv"
+            ],
+            "input_file_list": [],
+            "ra_column": "ra",
+            "dec_column": "dec",
+            "use_hipscat_index": false,
+            "sort_columns": null,
+            "constant_healpix_order": 1,
+            "lowest_healpix_order": 0,
+            "highest_healpix_order": 7,
+            "pixel_threshold": 1000000,
+            "mapping_healpix_order": 1,
+            "debug_stats_only": false,
+            "file_reader_info": {
+                "input_reader_type": "CsvReader",
+                "chunksize": 500000,
+                "schema_file": null,
+                "column_names": null,
+                "parquet_kwargs": null,
+                "kwargs": {}
+            }
+        }
+    }
+}
